Fix PyPI name collision to mcrypt-mass-crypto and migrate node bindings to koffi

+++ bindings/c_cpp/mcrypt.h
@@ -2,13 +2,14 @@
 #define MCRYPT_H
 
 #include <stdbool.h>
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 // Rust ನಿಂದ ಬರುವ ನೇರ ಫಂಕ್ಷನ್ಸ್
-const char* mcrypt_gensalt(unsigned int rounds, unsigned long salt_length);
+const char* mcrypt_gensalt(unsigned int rounds, size_t salt_length);
 const char* mcrypt_hash_with_salt(const char* password, const char* salt_prefix);
 bool mcrypt_verify(const char* password, const char* stored_hash);
 void mcrypt_free_string(char* s);
